Windows support
I made the program work on Windows 10 with Python 3.5 from windows Command Prompt (no Bash on Ubuntu on Windows stuff...). There are .bat files which should
do the job, please follow the instructions in README.md for installation. This is the first
commit with Windows support, bug fixes will follow.